relocates TransactionTracker to a new module.
